# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEPOT API",
        "contact": {
            "name": "oussema khlifi",
            "email": "khoussemaa@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/products/all": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Return all products",
                "operationId": "af88e80bb294939c1fc7e0231fba5ae0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/products/create": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Product store",
                "operationId": "5b4712b82366a108a816402b5d526fa5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "decimal"
                                    },
                                    "cost": {
                                        "type": "decimal"
                                    },
                                    "quantity": {
                                        "type": "decimal"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "expiredat": {
                                        "type": "date"
                                    },
                                    "depot_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/products/product/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Product show",
                "operationId": "57682f8a28d1c6b5e97b0d424134ede8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/products/update/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Product update",
                "operationId": "615bba49c084f6fc97e568daccec4639",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "decimal"
                                    },
                                    "cost": {
                                        "type": "decimal"
                                    },
                                    "quantity": {
                                        "type": "decimal"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "expiredat": {
                                        "type": "date"
                                    },
                                    "depot_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/products/delete/{id}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Summary",
                "operationId": "7efd851d8c2a224ae3945e8be4ef1691",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductSchema": {
                "title": "Product Model",
                "description": "Product Model",
                "properties": {
                    "name": {
                        "description": "product name",
                        "type": "string",
                        "example": "samasung A20"
                    },
                    "size": {
                        "description": "product size",
                        "type": "string",
                        "example": "20x40mm"
                    },
                    "weight": {
                        "description": "product weight",
                        "type": "decimal",
                        "example": "54g"
                    },
                    "cost": {
                        "description": "product cost",
                        "type": "decimal",
                        "example": "220"
                    },
                    "quantity": {
                        "description": "product quantity",
                        "type": "integer",
                        "example": "25"
                    },
                    "type": {
                        "description": "product type",
                        "type": "string",
                        "example": "smartphone"
                    },
                    "expiredat": {
                        "description": "product expiration date",
                        "type": "date",
                        "example": "2021/04/11"
                    },
                    "depot_id": {
                        "description": "product depot",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "query",
                "description": "depot id",
                "schema": {
                    "type": "number"
                }
            }
        }
    }
}